+++ src/custom-apps/components/WebhookDetailsPage/WebhookDetailsPage.tsx
@@ -26,6 +26,8 @@ import { parse, print } from "graphql";
 import React, { useEffect, useState } from "react";
 import { useIntl } from "react-intl";
 
+import PermissionAlert from "../PermissionAlert";
+import WebhookHeaders from "../WebhookHeaders";
 import WebhookSubscriptionQuery from "../WebhookSubscriptionQuery";
 import { getHeaderTitle } from "./messages";
 
@@ -37,6 +39,7 @@ export interface WebhookFormData {
   secretKey?: string;
   targetUrl: string;
   subscriptionQuery: string;
+  customHeaders: string;
 }
 
 export interface WebhookDetailsPageProps {
@@ -62,7 +65,7 @@ const WebhookDetailsPage: React.FC<WebhookDetailsPageProps> = ({
 
   let prettified: string;
   try {
-    prettified = print(parse(webhook?.subscriptionQuery));
+    prettified = print(parse(webhook?.subscriptionQuery || ""));
   } catch {
     prettified = webhook?.subscriptionQuery || "";
   }
@@ -75,6 +78,7 @@ const WebhookDetailsPage: React.FC<WebhookDetailsPageProps> = ({
     secretKey: webhook?.secretKey || "",
     targetUrl: webhook?.targetUrl || "",
     subscriptionQuery: prettified || "",
+    customHeaders: webhook?.customHeaders || "{}",
   };
 
   const backUrl = CustomAppUrls.resolveAppUrl(appId);
@@ -92,17 +96,18 @@ const WebhookDetailsPage: React.FC<WebhookDetailsPageProps> = ({
   return (
     <Form confirmLeave initial={initialForm} onSubmit={handleSubmit}>
       {({ data, submit, change }) => {
-        const handleSyncEventsSelect = createSyncEventsSelectHandler(
+        const handleSyncEventsSelect = createSyncEventsSelectHandler({
           change,
-          data.syncEvents,
-          setQuery,
-        );
-        const handleAsyncEventsSelect = createAsyncEventsSelectHandler(
-          change,
-          data.asyncEvents,
+          data,
           query,
           setQuery,
-        );
+        });
+        const handleAsyncEventsSelect = createAsyncEventsSelectHandler({
+          change,
+          data,
+          query,
+          setQuery,
+        });
 
         return (
           <DetailedContent useSingleColumn>
@@ -123,6 +128,7 @@ const WebhookDetailsPage: React.FC<WebhookDetailsPageProps> = ({
                 <FormSpacer />
                 <WebhookEvents
                   data={data}
+                  setQuery={setQuery}
                   onSyncEventChange={handleSyncEventsSelect}
                   onAsyncEventChange={handleAsyncEventsSelect}
                 />
@@ -131,6 +137,10 @@ const WebhookDetailsPage: React.FC<WebhookDetailsPageProps> = ({
                   setQuery={setQuery}
                   data={data}
                 />
+                <FormSpacer />
+                <PermissionAlert query={query} />
+                <FormSpacer />
+                <WebhookHeaders data={data} onChange={change} />
               </Box>
             </Content>
             <Savebar