+++ src/graphql/hooks.generated.ts
@@ -2841,6 +2841,7 @@ export const WebhookDetailsFragmentDoc = gql`
   secretKey
   targetUrl
   subscriptionQuery
+  customHeaders
 }
     ${WebhookFragmentDoc}`;
 export const AppCreateDocument = gql`