+++ src/productTypes/components/ProductTypeListPage/ProductTypeListPage.tsx
@@ -64,8 +64,8 @@ const ProductTypeListPage: React.FC<ProductTypeListPageProps> = ({
           data-test-id="add-product-type"
         >
           <FormattedMessage
-            id="QY7FSs"
-            defaultMessage="create product type"
+            id="gksZwp"
+            defaultMessage="Create product type"
             description="button"
           />
         </Button>